move and update dialogs
