+++ src/internal/components/index.ts
@@ -1,5 +1,6 @@
 export * from './common';
 export * from './dataset';
+export * from './login';
 // eslint-disable-next-line import/export
 export * from './session';
 export * from './tours';
